formatting changes

+++ scripts/common_functions.sh
@@ -1,40 +1,42 @@
 #!bin/bash
 
-
 unzipKit(){
-deleteFileAfterUnzip=false
-#download if URI is provided
-if [ "$KIT_URL" != "" ]; then
-  deleteFileAfterUnzip=true
-  buildname=$(basename $KIT_URL)
-  if curl --output /dev/null --silent --head --fail $KIT_URL
+if [ -z "$(ls -A /opt/pega/kit/scripts)" -a -z "$(ls -A /opt/pega/kit/archives)" -a -z "$(ls -A /opt/pega/kit/rules)" ]
   then
-    echo "Downloading kit from url"
-    curl -ksSL -o ${kit_root}/$buildname ${KIT_URL}
-  else
-    echo "Unable to download kit from url"
-  fi
+    deleteFileAfterUnzip=false
+    #download if URI is provided
+    if [ "$KIT_URL" != "" ]; then
+      deleteFileAfterUnzip=true
+      buildname=$(basename $KIT_URL)
+      if curl --output /dev/null --silent --head --fail $KIT_URL
+      then
+        echo "Downloading kit from url"
+        curl -ksSL -o ${kit_root}/$buildname ${KIT_URL}
+      else
+        echo "Unable to download kit from url"
+      fi
+    fi
+
+    #check the no of zip files
+      no_of_zip_files=$(ls -lr ${kit_root}/*.zip 2> /dev/null| wc -l)
+      if [ "$no_of_zip_files" == 0 ]
+      then
+    	echo "The distribution kit is neither mounted nor downloaded through Kit url '"$KIT_URL"'. Please provide a valid distribution image"
+    	exit 1;
+      elif [ "$no_of_zip_files" -gt 1 ]
+      then
+    	echo "/opt/pega/kit folder should contain only one distribution image in zip format"
+        exit 1     
+      else
+         unzip -o ${kit_root}/*.zip -d ${kit_root}
+      fi
+
+      #delete the file if URI is provided
+      if $deleteFileAfterUnzip ; then
+        echo "Deleting downloaded kit zip file."
+        rm ${kit_root}/$buildname
+      fi
 fi
-
-#check the no of zip files
-  no_of_zip_files=$(ls -lr ${kit_root}/*.zip 2> /dev/null| wc -l)
-  if [ "$no_of_zip_files" == 0 ]
-  then
-	echo "The distribution kit is neither mounted nor downloaded through Kit url '"$KIT_URL"'. Please provide a valid distribution image"
-	exit 1;
-  elif [ "$no_of_zip_files" -gt 1 ]
-  then
-	echo "/opt/pega/kit folder should contain only one distribution image in zip format"
-    exit 1     
-  else
-     unzip -o ${kit_root}/*.zip -d ${kit_root}
-  fi
-
-  #delete the file if URI is provided
-  if $deleteFileAfterUnzip ; then
-    echo "Deleting downloaded kit zip file."
-    rm ${kit_root}/$buildname
-  fi
 }
 
 
@@ -81,35 +83,67 @@ export DRIVER_JAR_PATH=$DRIVER_JAR_PATH
 
 # Read Secrets
 readSecrets() {
-    if [ -e "$db_username_file" ]; then
-        SECRET_DB_USERNAME=$(<${db_username_fil e})
-    else
-        SECRET_DB_USERNAME=${DB_USERNAME}
-    fi
+  if [ -e "$db_username_file" ]; then
+      SECRET_DB_USERNAME=$(<${db_username_fil e})
+  else
+      SECRET_DB_USERNAME=${DB_USERNAME}
+  fi
 
-    if [ -e "$db_password_file" ]; then
-        SECRET_DB_PASSWORD=$(<${db_password_file})
-    else
-        SECRET_DB_PASSWORD=${DB_PASSWORD}
-    fi
+  if [ -e "$db_password_file" ]; then
+      SECRET_DB_PASSWORD=$(<${db_password_file})
+  else
+      SECRET_DB_PASSWORD=${DB_PASSWORD}
+  fi
 
-    if { [ "$SECRET_DB_USERNAME" == "" ] || [ "$SECRET_DB_PASSWORD" == "" ] ;} && [ ! -e "$setupDatabase_properties" ]; then
-        echo "DB_USERNAME and DB_PASSWORD must be specified.";
-    exit 1
-    fi
-    # This needs to be exported for dockerize to correctly replace USERNAME, PASSWORD in template files
-    export DB_USERNAME=$SECRET_DB_USERNAME
-    export DB_PASSWORD=$SECRET_DB_PASSWORD
+  if { [ "$SECRET_DB_USERNAME" == "" ] || [ "$SECRET_DB_PASSWORD" == "" ] ;} && [ ! -e "$setupDatabase_properties" ]; then
+      echo "DB_USERNAME and DB_PASSWORD must be specified.";
+  exit 1
+  fi
+  # This needs to be exported for dockerize to correctly replace USERNAME, PASSWORD in template files
+  export DB_USERNAME=$SECRET_DB_USERNAME
+  export DB_PASSWORD=$SECRET_DB_PASSWORD
 }
 
 # Mount prlog4j2 provided
 mount_log4j2() {
-if [ -e "$prlog4j2" ]; then
-    echo "Loading prlog4j2 from ${prlog4j2}...";
-    cp "$prlog4j2" ${scripts_root}/config
-else
-    echo "No prlog4j2 was specified in ${prlog4j2}.  Using defaults."
-fi
+  if [ -e "$prlog4j2" ]; then
+      echo "Loading prlog4j2 from ${prlog4j2}...";
+      cp "$prlog4j2" ${scripts_root}/config
+  else
+      echo "No prlog4j2 was specified in ${prlog4j2}.  Using defaults."
+  fi
+}
+
+# Mount provided database conf files 
+mount_dbconf() {
+  if [ -e "$mssql_conf" ] && [ "$DB_TYPE" == "mssql" ]; then
+      echo "Loading mssql.conf from ${mssql_conf}...";
+      cp "$mssql_conf" "${scripts_root}/config/mssql/mssql.conf"
+
+  elif [ -e "$postgres_conf" ] && [ "$DB_TYPE" == "postgres" ]; then
+      echo "Loading postgres.conf from ${postgres_conf}...";
+      cp "$postgres_conf" "${scripts_root}/config/postgres/postgres.conf"
+
+  elif [ -e "$oracledate_conf" ] && [ "$DB_TYPE" == "oracledate" ]; then
+      echo "Loading oracledate.conf from ${oracledate_conf}...";
+      cp "$oracledate_conf" "${scripts_root}/config/oracledate/oracledate.conf"
+
+  elif [ -e "$udb_conf" ] && [ "$DB_TYPE" == "udb" ]; then
+      echo "Loading udb.conf from ${udb_conf}...";
+      cp "$udb_conf" "${scripts_root}/config/udb/udb.conf"
+
+  elif [ -e "$db2zos_conf" ] && [ "$DB_TYPE" == "db2zos" ]; then
+      echo "Loading db2zos.conf from ${db2zos_conf}...";
+      cp "$db2zos_conf" "${scripts_root}/config/db2zos/db2zos.conf"
+  fi 
+}
+
+mount_files()
+{
+
+  # mount these files before dockerizing prconfig, prbootstrap, setupdatabase
+  mount_log4j2
+  mount_dbconf
 }
 
 # Dockerize setupDatabase.properties from template
@@ -136,54 +170,36 @@ dockerizeMigrateSystemProperties() {
     /bin/dockerize -template ${config_root}/migrateSystem.properties.tmpl:${scripts_root}/migrateSystem.properties
 }
 
-# Mount provided database conf files 
-mount_dbconf() {
-if [ -e "$mssql_conf" ] && [ "$DB_TYPE" == "mssql" ]; then
-    echo "Loading mssql.conf from ${mssql_conf}...";
-    cp "$mssql_conf" "${scripts_root}/config/mssql/mssql.conf"
-
-elif [ -e "$postgres_conf" ] && [ "$DB_TYPE" == "postgres" ]; then
-    echo "Loading postgres.conf from ${postgres_conf}...";
-    cp "$postgres_conf" "${scripts_root}/config/postgres/postgres.conf"
-
-elif [ -e "$oracledate_conf" ] && [ "$DB_TYPE" == "oracledate" ]; then
-    echo "Loading oracledate.conf from ${oracledate_conf}...";
-    cp "$oracledate_conf" "${scripts_root}/config/oracledate/oracledate.conf"
-
-elif [ -e "$udb_conf" ] && [ "$DB_TYPE" == "udb" ]; then
-    echo "Loading udb.conf from ${udb_conf}...";
-    cp "$udb_conf" "${scripts_root}/config/udb/udb.conf"
-
-elif [ -e "$db2zos_conf" ] && [ "$DB_TYPE" == "db2zos" ]; then
-    echo "Loading db2zos.conf from ${db2zos_conf}...";
-    cp "$db2zos_conf" "${scripts_root}/config/db2zos/db2zos.conf"
-fi 
+dockerizeFiles()
+{
+  dockerizePrconfig
+  dockerizePrbootstrap
+  dockerizeSetupdatabase
 }
-
 
 
 initializeSchemas() 
 {
- if [ "$DATA_SCHEMA" == '' ]; then
+  if [ "$DATA_SCHEMA" == '' ]; then
     export DATA_SCHEMA=$RULES_SCHEMA
- fi
- if [ "$CUSTOMERDATA_SCHEMA" == '' ]; then
+  fi
+  if [ "$CUSTOMERDATA_SCHEMA" == '' ]; then
     export CUSTOMERDATA_SCHEMA=$DATA_SCHEMA
- fi
-ACTUAL_RULES_SCHEMA=$RULES_SCHEMA;
-ACTUAL_DATA_SCHEMA=$DATA_SCHEMA;
-ACTUAL_CUSTOMERDATA_SCHEMA=$CUSTOMERDATA_SCHEMA;
+  fi
+  ACTUAL_RULES_SCHEMA=$RULES_SCHEMA;
+  ACTUAL_DATA_SCHEMA=$DATA_SCHEMA;
+  ACTUAL_CUSTOMERDATA_SCHEMA=$CUSTOMERDATA_SCHEMA;
 
 }
 
 isActionValid() {
- echo "Checking valid action"
- VALID_ACTIONS="install, upgrade, install-deploy, upgrade-deploy, pre-upgrade, post-upgrade"
- if [ -n "`echo $VALID_ACTIONS | xargs -d ","| xargs -n1| grep -e \"^$ACTION$\"`" ]; then
-  echo "Action selected is '"$ACTION"'";
- else
-  echo "Invalid action '"$ACTION"' passed.";
-  echo "Valid actions are : " $VALID_ACTIONS;
-  exit 1;
- fi
+  echo "Checking valid action"
+  VALID_ACTIONS="install, upgrade, install-deploy, upgrade-deploy, pre-upgrade, post-upgrade"
+  if [ -n "`echo $VALID_ACTIONS | xargs -d ","| xargs -n1| grep -e \"^$ACTION$\"`" ]; then
+    echo "Action selected is '"$ACTION"'";
+  else
+    echo "Invalid action '"$ACTION"' passed.";
+    echo "Valid actions are : " $VALID_ACTIONS;
+    exit 1;
+  fi
 }
